requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dateparser==1.2.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
numpy==1.26.4
pandas==2.2.0
pyarrow==15.0.0
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.0
websocket-client==1.7.0
websockets==12.0
yarl==1.9.4
